# DefaultPromise MicroBenchmark