Motivation:
There is no microbench for DefaultPromise.

Modifications:
- Create a benchmark which covers a few mainstream use cases of DefaultPromise

Result:
JMH microbench exists for DefaultPromise