@Scottmitch 2016-05-06
@ejona86 @nmittler @buchgr - Would you mind giving me your input on the "validity" of this benchmark ... and plz make suggestions if you have a better / different approach