@Scottmitch 2016-05-06
I created this to benchmark the impacts of https://github.com/netty/netty/pull/5221 ... the performance looks similar but that could be bcz the benchmark isn't really measuring things in a meaningful way...

```
current 4.1 HEAD
Benchmark                                                   (lateListeners)  (numListeners)  (numPromises)   Mode  Cnt        Score        Error  Units
DefaultPromiseBenchmark.testInEventLoop                                   0               1        1000000  thrpt    5        0.999 ±      0.001  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 0               1        1000000  thrpt    5  1997431.346 ±    721.472  ops/s
DefaultPromiseBenchmark.testInEventLoop                                   0               4        1000000  thrpt    5        0.999 ±      0.001  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 0               4        1000000  thrpt    5  7988008.447 ±   6771.461  ops/s
DefaultPromiseBenchmark.testInEventLoop                                   4               1        1000000  thrpt    5        0.998 ±      0.002  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 4               1        1000000  thrpt    5  1996448.985 ±   3835.691  ops/s
DefaultPromiseBenchmark.testInEventLoop                                   4               4        1000000  thrpt    5        0.998 ±      0.001  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 4               4        1000000  thrpt    5  7986887.112 ±  11432.078  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  0               1        1000000  thrpt    5        0.704 ±      0.014  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                0               1        1000000  thrpt    5  1408743.076 ±  27481.636  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  0               4        1000000  thrpt    5        0.514 ±      0.022  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                0               4        1000000  thrpt    5  4110335.690 ± 178393.550  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  4               1        1000000  thrpt    5        0.685 ±      0.110  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                4               1        1000000  thrpt    5  1369759.856 ± 219828.634  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  4               4        1000000  thrpt    5        0.457 ±      0.016  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                4               4        1000000  thrpt    5  3655099.832 ± 124950.326  ops/s

new PR
Benchmark                                                   (lateListeners)  (numListeners)  (numPromises)   Mode  Cnt        Score        Error  Units
DefaultPromiseBenchmark.testInEventLoop                                   0               1        1000000  thrpt    5        0.999 ±      0.001  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 0               1        1000000  thrpt    5  1997308.945 ±    263.894  ops/s
DefaultPromiseBenchmark.testInEventLoop                                   0               4        1000000  thrpt    5        0.999 ±      0.001  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 0               4        1000000  thrpt    5  7988727.222 ±   7406.745  ops/s
DefaultPromiseBenchmark.testInEventLoop                                   4               1        1000000  thrpt    5        0.999 ±      0.001  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 4               1        1000000  thrpt    5  1997270.605 ±   1578.901  ops/s
DefaultPromiseBenchmark.testInEventLoop                                   4               4        1000000  thrpt    5        0.999 ±      0.001  ops/s
DefaultPromiseBenchmark.testInEventLoop:messagesPerSecond                 4               4        1000000  thrpt    5  7988485.416 ±   9247.465  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  0               1        1000000  thrpt    5        0.680 ±      0.086  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                0               1        1000000  thrpt    5  1360177.708 ± 171723.573  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  0               4        1000000  thrpt    5        0.606 ±      0.012  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                0               4        1000000  thrpt    5  4851124.127 ±  99226.553  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  4               1        1000000  thrpt    5        0.695 ±      0.064  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                4               1        1000000  thrpt    5  1390348.528 ± 127220.929  ops/s
DefaultPromiseBenchmark.testOutEventLoop                                  4               4        1000000  thrpt    5        0.602 ±      0.019  ops/s
DefaultPromiseBenchmark.testOutEventLoop:messagesPerSecond                4               4        1000000  thrpt    5  4816534.137 ± 151259.958  ops/s
```