@Scottmitch 2016-05-09
@normanmaurer @nmittler - PTAL. 1 open question https://github.com/netty/netty/pull/5222#discussion_r62549140